+++ fast_rx/lib/fast_rx.dart
@@ -15,7 +15,9 @@ export 'src/rx/object/rx_set.dart';
 
 // rx/primitive
 export 'src/rx/primitive/rx_bool.dart';
+export 'src/rx/primitive/rx_date_time.dart';
 export 'src/rx/primitive/rx_double.dart';
+export 'src/rx/primitive/rx_duration.dart';
 export 'src/rx/primitive/rx_int.dart';
 export 'src/rx/primitive/rx_string.dart';